change trafic field

+++ panel/models.py
@@ -24,7 +24,7 @@ class Profile_Extended(models.Model):
     key_link = models.ForeignKey(Key_Link_List, on_delete=models.CASCADE)
     domanin_rank = models.CharField(max_length=500, default="")
     domanin_auth = models.CharField(max_length=500, default="")
-    traffic = models.IntegerField(blank=False)
+    traffic = models.FloatField(blank=False)
     spam_score = models.IntegerField(blank=False)
     existing_cost = models.IntegerField(blank=False)
     new_cost = models.IntegerField(blank=False)
